+++ NINA.Plugin.SeeDrift/Properties/AssemblyInfo.cs
@@ -1,6 +1,9 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 using System.Windows;
+
+[assembly: SupportedOSPlatform("windows")]
 
 [assembly: AssemblyTitle("SeeDrift")]
 [assembly: AssemblyMetadata("MinimumApplicationVersion", "3.2.0.9001")]
@@ -13,8 +16,8 @@ using System.Windows;
 When log lines correlate between exposures, dither and center-after-drift triggers appear in the HTML report.")]
 [assembly: AssemblyCompany("Carl Stovell")]
 [assembly: AssemblyProduct("NINA.Plugin.SeeDrift")]
-[assembly: AssemblyVersion("0.7.9.0")]
-[assembly: AssemblyFileVersion("0.7.9.0")]
+[assembly: AssemblyVersion("0.7.10.0")]
+[assembly: AssemblyFileVersion("0.7.10.0")]
 [assembly: AssemblyMetadata("id", "E8D4C2F1-6B7A-4E9D-8C3F-1A2B5D7E9F0C")]
 [assembly: Guid("E8D4C2F1-6B7A-4E9D-8C3F-1A2B5D7E9F0C")]
 [assembly: ThemeInfo(